internal: promote toFloat64 from ddtrace/tracer
